src/main/java/ru/codeseeker/persistence/model/Ticket.java
package ru.codeseeker.persistence.model;

import org.hibernate.annotations.Type;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Table;
import java.math.BigDecimal;

@Entity
@Table(name = "ticket")
public class Ticket extends BaseEntity {
    @Column
    private String name;

    @Column
    private BigDecimal cost;

    @Type(type = "jsonb")
    @Column(columnDefinition = "jsonb")
    private Location location;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public BigDecimal getCost() {
        return cost;
    }

    public void setCost(BigDecimal cost) {
        this.cost = cost;
    }

    public Location getLocation() {
        return location;
    }

    public void setLocation(Location location) {
        this.location = location;
    }
}
